added mutations

+++ server/schemas/resolvers.js
@@ -180,18 +180,18 @@ const resolvers = {
             throw new AuthenticationError("You need to be logged in!");
         },
         deleteMusic: async (parent, args, context) => {
-            // if (context.user) {
+            if (context.user) {
                 await Music.deleteOne( {_id: args._id})                    
                 
                 return args._id;
-            // }
-            // throw new AuthenticationError("You need to be logged in!");
+            }
+            throw new AuthenticationError("You need to be logged in!");
         },
         editMusic: async (parent, args, context) => {
-            // if (context.user) {
+            if (context.user) {
                 return await Music.findByIdAndUpdate(args.updatedMusic._id, args.updatedMusic, {new: true})                    
-            // }
-            // throw new AuthenticationError("You need to be logged in!");
+            }
+            throw new AuthenticationError("You need to be logged in!");
         }
     }
 };
